refactor(vscode): respect sendMethod for multi-line inspect targets
- Extract send_code, get_send_options, and the bracketed-paste/tempfile
  transports from commands.ts into a new send-code.ts so inspect-commands
  can reuse them.
- Route inspection commands through send_code only when the wrapped
  expression is multi-line (the user explicitly selected a multi-line
  expression to inspect). Single-line wrapped exprs go straight to the
  terminal so `tempfile` mode doesn't spool a one-liner like nrow(x) to
  disk.
